requirements.txt
Django==5.1.3
djangorestframework==3.14.0
channels==4.0.0
channels_redis==4.0.0
psycopg2-binary==2.9.6
